Select a unit on left click even if one's selected
It's annoying to have to cancel a selection in order to select another unit. Just left click another unit to select it even if not in "selection" state. This doesn't conflict with movement and attacks, because movements only work on left clicks on empty spaces and attacks only work on right clicks on other units.