Fixed valid check in bfs, movement almost correct.
At this point the bfs is returning the right results. The only problem
is with react.

There needs to be a good place to perform the bfs and store away the
results. I think I need to add the bfs results to state after a move.

The main problem I'm having is that the cell glows linger. I need to
debug why thats happening, but I think its because no state change has
happened for *those* cells. But its need to know that they've no longer
movable.